app/routers/ai_chat.py
import os
from fastapi import APIRouter, HTTPException
from dotenv import load_dotenv
from google import genai
from ..schemas import ChatRequest

# Load environment variables
load_dotenv()

router = APIRouter(prefix="/ai", tags=["AI"])

# Get API Key
api_key = os.getenv("GEMINI_API_KEY")

if not api_key:
    raise ValueError("GEMINI_API_KEY not found in .env file")

# Initialize Gemini Client
client = genai.Client(api_key=api_key)

# Medical safety system prompt
SYSTEM_PROMPT = """
You are Healthify AI, a healthcare assistant.

Rules:
- Provide only general medical guidance.
- Do NOT provide prescriptions or exact dosages.
- Do NOT diagnose diseases.
- Always end response with:
  'Consult a certified doctor for proper diagnosis.'
"""

@router.post("/chat")
def chat(request: ChatRequest):
    try:
        full_prompt = f"{SYSTEM_PROMPT}\n\nUser: {request.message}"

        response = client.models.generate_content(
            model="gemini-2.5-flash",
            contents=full_prompt,
        )

        return {
            "response": response.text
        }

    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))