requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
google-genai